vim: pylint syntax highlighter
I think this could be vastly improved so I'm going to just leave this in
the dotfile repo. Something I need to figure out which also applies to
the ESLint highlighter is how to get consistent greens, yellows, and
reds that belong to the color scheme set by the user. There doesn't seem
to be a mechanic for this in Vim, and at least gruxbox's strategy is to
recolor syntax match groups to better fit their intent. That does mean
they have a huge color scheme file though...surely there's a better
strategy?